Deleted SDL 1.2 and boost. Hello SDL2 and c++11.

+++ game.h
@@ -7,24 +7,24 @@ namespace Guy {
 class Game
 {
 public:
-    static Game* ptrInstance();
+	static Game* ptrInstance();
 
 	Game();
-    virtual ~Game();
+	virtual ~Game();
 
 	virtual void init();
 	virtual void load();
 
-	virtual void update(float deltaTime) = 0;
+	virtual void update(double deltaTime) = 0;
 	virtual void draw() = 0;
 
 	virtual void unload();
 
 	virtual const char *name();
-    virtual const char *version();
+	virtual const char *version();
 
 private:
-    static Game *s_game;
+	static Game *s_game;
 };
 
 } //namespace Guy
