+++ graphics/tools/frameshud.h
@@ -11,18 +11,18 @@ class FramesHUD
 	FramesHUD();
 	bool loadFont(const char *filename);
 	void setColor(const rgba &color);
-	void setDisplayTime(float time);
+	void setDisplayTime(double time);
 
-	void update(float deltaTime);
+	void update(double deltaTime);
 	void draw();
 	void reset();
 
 protected:
 	Font m_font;
 	rgba m_color;
-	float m_displayTime, m_displayTimeInv;
-	float m_timeCount;
-	float m_framesToDisplay;
+	double m_displayTime, m_displayTimeInv;
+	double m_timeCount;
+	double m_framesToDisplay;
 	unsigned short m_framesCount;
 	bool m_isFontLoaded;
 };
